refactor(data-table): change a type export
